[FLINK-32074][checkpoint] Merge file across checkpoints
