+++ flink-runtime/src/main/java/org/apache/flink/runtime/checkpoint/filemerging/FileMergingSnapshotManagerBuilder.java
@@ -29,6 +29,9 @@ public class FileMergingSnapshotManagerBuilder {
     /** The id for identifying a {@link FileMergingSnapshotManager}. */
     private final String id;
 
+    /** The file merging type. */
+    private final FileMergingType fileMergingType;
+
     /** Max size for a file. TODO: Make it configurable. */
     private long maxFileSize = 32 * 1024 * 1024;
 
@@ -42,8 +45,9 @@ public class FileMergingSnapshotManagerBuilder {
      *
      * @param id the id of the manager.
      */
-    public FileMergingSnapshotManagerBuilder(String id) {
+    public FileMergingSnapshotManagerBuilder(String id, FileMergingType type) {
         this.id = id;
+        this.fileMergingType = type;
     }
 
     /** Set the max file size. */
@@ -71,13 +75,27 @@ public class FileMergingSnapshotManagerBuilder {
     /**
      * Create file-merging snapshot manager based on configuration.
      *
-     * <p>TODO (FLINK-32074): Support another type of FileMergingSnapshotManager that merges files
-     * across different checkpoints.
-     *
      * @return the created manager.
      */
     public FileMergingSnapshotManager build() {
-        return new WithinCheckpointFileMergingSnapshotManager(
-                id, maxFileSize, filePoolType, ioExecutor == null ? Runnable::run : ioExecutor);
+        switch (fileMergingType) {
+            case MERGE_WITHIN_CHECKPOINT:
+                return new WithinCheckpointFileMergingSnapshotManager(
+                        id,
+                        maxFileSize,
+                        filePoolType,
+                        ioExecutor == null ? Runnable::run : ioExecutor);
+            case MERGE_ACROSS_CHECKPOINT:
+                return new AcrossCheckpointFileMergingSnapshotManager(
+                        id,
+                        maxFileSize,
+                        filePoolType,
+                        ioExecutor == null ? Runnable::run : ioExecutor);
+            default:
+                throw new UnsupportedOperationException(
+                        String.format(
+                                "Unsupported type %s when creating file merging manager",
+                                fileMergingType));
+        }
     }
 }
